+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [2.0.1]
 ### Updated
 - Correct data extraction validator to accept columns: all
+- Add support  for SQL queries on client
+- Adapt test queries to the changes on SlicingDice API
 
 ## [2.0.0]
 ### Updated